% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_per_celltype.R
\name{description_per_celltype}
\alias{description_per_celltype}
\title{Description of various statistics for each celltype in the Seurat object}
\usage{
description_per_celltype(
  seurat_object,
  condition_to_compare,
  condition_ident,
  celltype_ident,
  sample_ident,
  celltypes
)
}
\arguments{
\item{seurat_object}{Seurat single-cell RNA sequencing object}
}
\description{
Description of various statistics for each celltype in the Seurat object
}
