% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_prop.R
\name{sc_prop}
\alias{sc_prop}
\title{sc_prop}
\usage{
sc_prop(
  seurat_object,
  conditions,
  celltypes,
  outdir,
  condition_ident = "condition",
  sample_ident = "sample",
  celltype_ident = "celltype",
  cols = NULL
)
}
\arguments{
\item{seurat_object}{Seurat single-cell RNA sequencing object}

\item{conditions}{Vector of conditions to compare. All conditions must be present in the Seurat object. Order in which conditions are given determines order in the final figure}

\item{celltypes}{Vector of celltypes to compare in the Seurat object}

\item{outdir}{Output directory. Will raise error if does not exist. Default: current directory}

\item{condition_ident}{Name of the metadata column containing the condition. Default: condition.}

\item{sample_ident}{Name of the metadata column containing the sample information. Default: sample}

\item{celltype_ident}{Name of the metadata column (Seurat Ident) containing the celltypes. Default: celltype}
}
\value{
Output consists of:
\enumerate{
\item dirichlet_plot.pdf: Ggplot2 visualization of the cell proportions per condition
\item dirichlet_regression_results.csv: Results for all comparisons made
\item dirichlet_regression_results_sig.csv: Significant results if any
}
}
\description{
Dirichlet regression to test for differences in cell proportions between conditions in single-cell RNA sequencing data
}
